fix: 삭제 엔드포인트 수정

+++ daker/src/main/java/com/daker/service/TeamService.java
@@ -363,8 +363,11 @@ public class TeamService {
     }
 
     // 모집글 삭제
-    public void deleteArticle(Long userId, Long articleId, ArticleRequestDTO.ArticleIdDTO request){
-        Article article = articleRepository.findById(articleId).orElseThrow(() -> new ApiException(NOT_FOUND_404));
+    public void deleteArticle(Long userId, ArticleRequestDTO.ArticleIdDTO request){
+        User user = userRepository.findById(userId).orElseThrow(() -> new ApiException(USER_NOT_FOUND_404));
+        Article article = articleRepository.findById(request.getArticleId()).orElseThrow(() -> new ApiException(NOT_FOUND_404));
+        if(article.getWriter().getId() != user.getId()) throw new ApiException(UNAUTHORIZED_401);
+
         articleRepository.delete(article);
     }
